+++ android/app/src/main/AndroidManifest.xml
@@ -71,7 +71,6 @@
         </activity>
         <activity android:name="com.facebook.react.devsupport.DevSettingsActivity"/>
         <service android:name="im.status.ethereum.pushnotifications.ForegroundService"></service>
-        <service android:name="im.status.ethereum.module.LocalNotificationsService" />
         <provider
             android:name="androidx.core.content.FileProvider"
             android:authorities="${applicationId}.provider"
