--- modules/react-native-status/android/src/main/java/im/status/ethereum/module/LocalNotificationsService.java
@@ -1,20 +1,0 @@
-package im.status.ethereum.module;
-
-import android.content.Intent;
-import android.os.Bundle;
-import com.facebook.react.HeadlessJsTaskService;
-import com.facebook.react.bridge.Arguments;
-import com.facebook.react.jstasks.HeadlessJsTaskConfig;
-import javax.annotation.Nullable;
-
-public class LocalNotificationsService extends HeadlessJsTaskService {
-
-    @Override
-    protected @Nullable HeadlessJsTaskConfig getTaskConfig(Intent intent) {
-        Bundle extras = intent.getExtras();
-        if (extras != null) {
-            return new HeadlessJsTaskConfig("LocalNotifications", Arguments.fromBundle(extras), 60000, true);
-        }
-        return null;
-    }
-}
