+++ modules/react-native-status/android/src/main/java/im/status/ethereum/pushnotifications/ForegroundService.java
@@ -55,7 +55,7 @@ public class ForegroundService extends Service {
         intent.setAction(Intent.ACTION_VIEW);
         intent.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK | Intent.FLAG_ACTIVITY_CLEAR_TOP);
         PendingIntent pendingIntent = PendingIntent.getActivity(this, 0, intent, 0);
-        Intent stopIntent = new Intent(NewMessageSignalHandler.ACTION_TAP_STOP);
+        Intent stopIntent = new Intent(PushNotificationHelper.ACTION_TAP_STOP);
         PendingIntent stopPendingIntent = PendingIntent.getBroadcast(context, 0, stopIntent, PendingIntent.FLAG_CANCEL_CURRENT);
 
         String content = context.getResources().getString(R.string.keep_status_running);
