+++ status-go-version.json
@@ -2,7 +2,7 @@
     "_comment": "DO NOT EDIT THIS FILE BY HAND. USE 'scripts/update-status-go.sh <tag>' instead",
     "owner": "status-im",
     "repo": "status-go",
-    "version": "v0.74.1",
-    "commit-sha1": "5a76e93063e3b9ef2ded83a5f545f7149401e8f0",
-    "src-sha256": "0581048fvqg0fxv03pm9a0drfiq32slxliwamnch3052avpdkycj"
+    "version": "v0.74.2",
+    "commit-sha1": "1724ecffa178e9bcb04ae6d9d95d276c02878b6f",
+    "src-sha256": "1kn6996brpva1bdj8bv1mscwx17hynqk8xb667ai6qpdriscr22d"
 }
